+++ app/components/FireBaseMessaging.js
@@ -561,7 +561,6 @@ class FireBaseMessaging extends Component {
     }
 
     if (stateChanged && isIOS) {
-      this.updateApplicationIconBadgeNumber();
       updateBadgeNumber(this.notificationsCount);
     }